msm_fb: Discontinue use of clk_set_min_rate()
This proprietary API will soon be removed. Replace use of it with a
combination of clk_round_rate() and clk_set_rate() where appropriate.

Calls to clk_set_min_rate() with a rate of 0Hz followed immediately
by a clk_disable() served no purpose and are removed entirely.

Change-Id: I56d680ee36cc6292dcc436eeec7acdc57f3c243a
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/clock-7x30.c
	arch/arm/mach-msm/clock-pcom-lookup.c

Conflicts:

	arch/arm/mach-msm/clock-7x30.c
	drivers/video/msm/mdp.c